tests/justhtml-tests/head_noscript_recovery.dat

#data
<head><noscript><!DOCTYPE html></noscript>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,31): unexpected-doctype

#script-off
#document
| <html>
|   <head>
|     <noscript>
|   <body>

#data
<head><noscript></p></noscript>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,20): unexpected-end-tag

#script-off
#document
| <html>
|   <head>
|     <noscript>
|   <body>

#data
</head><noscript>
#errors
(1,1): expected-doctype-but-got-chars
(1,1): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <noscript>


#data
<head><noscript><head class="foo"><!--foo--></noscript>
#errors
(1,6): expected-doctype-but-got-start-tag
#script-off
#document
| <html>
|   <head>
|     <noscript>
|       <!-- foo -->
|   <body>

#data
<noscript><script>
#errors
#script-off
#document
| <html>
|   <head>
|     <noscript>
|     <script>
|   <body>
